Prepare release of version v2.0.0 (#146)
* Update Spack-Repo to account for MFEM 4.9 + Hypre 3.1.0 and prepare release of v2.0.0

* Update version number in CMakeLists.txt

* Update Doxyfile.cmake

* Debug recommandations

* Add student contributor

+++ kernel/include/Integrators/RobinNLFormIntegrator.tpp
@@ -157,7 +157,7 @@ void RobinNLFormIntegrator<VARS>::AssembleElementVector(
     *elvect[blk] = 0.;
 
     // Add Robin on the right block
-    if (blk == blk_) {
+    if (blk == static_cast<int>(blk_)) {
       const mfem::IntegrationRule* ir =
           &mfem::IntRules.Get(el[blk]->GetGeomType(), 2 * el[blk]->GetOrder() + Tr.OrderW());
 
@@ -225,7 +225,7 @@ void RobinNLFormIntegrator<VARS>::AssembleElementGrad(
     *elmat(blk, blk) = 0.0;
 
     // Add Robin on the right block
-    if (blk == blk_) {
+    if (blk == static_cast<int>(blk_)) {
       const mfem::IntegrationRule* ir =
           &mfem::IntRules.Get(el[blk]->GetGeomType(), 2 * el[blk]->GetOrder() + Tr.OrderW());
